requirements.txt
asgiref==3.8.1
autopep8==2.1.0
Django==4.2.11
djangorestframework==3.15.1
pycodestyle==2.11.1
pytz==2024.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
